move PackageCacheRecord to records.py
